Space optimized dynamic programming approach
First attempt:
https://www.geeksforgeeks.org/0-1-knapsack-problem-dp-10/
The knapsack is similar to the pizza problem, but a bit more complex, once you have weights and values for each item.
In our pizza, we only have weights. 
I first tested the knapsack solution applied to our pizza problem, just by making the values vector equal to the weights vector.
The solution worked but had memory issues.

Second attempt:
Then I tried doing the same in this solution but encountered the memory issues as well.
https://www.geeksforgeeks.org/space-optimized-dp-solution-0-1-knapsack-problem/

Obviously, the best approach is to adapt the code, once our problem is simpler and don't require the consideration of a value vector.